+++ R/prepare_wham_om_input.R
@@ -494,7 +494,9 @@ without changing ASAP file, specify M$initial_means.")
   data$simulate_data = rep(1,3) #simulate data types (catch, indices, Ecov)
   data$simulate_period = rep(1,2) #simulate above items for (model years, projection years)
   data$percentSPR = 40 #percentage of unfished SSB/R to use for SPR-based reference points
-  data$XSPR_R_opt = 3 #1(3): use annual R estimates(predictions) for annual SSB_XSPR, 2(4): use average R estimates(predictions). See next line for years to average over.
+  data$percentFXSPR = 100 # percent of F_XSPR to use for calculating catch in projections
+  data$percentFMSY = 100 # percent of F_XSPR to use for calculating catch in projections
+	data$XSPR_R_opt = 3 #1(3): use annual R estimates(predictions) for annual SSB_XSPR, 2(4): use average R estimates(predictions). See next line for years to average over.
   data$XSPR_R_avg_yrs = 1:data$n_years_model #model year indices (TMB, starts @ 0) to use for averaging recruitment when defining SSB_XSPR (if XSPR_R_opt = 2,4)
   model_years <- basic_info$modyears
 
@@ -841,6 +843,11 @@ Ex: ",ecov$label[i]," in ",years[1]," affects ", c('recruitment','M')[data$Ecov_
   data$proj_F_opt <- 0
   data$proj_Fcatch <- 0
   data$proj_M_opt <- 0
+  data$logR_mean <- 0 # only used for SCAA projections
+  data$logR_sd <- 0 # only used for SCAA projections
+  data$FXSPR_init = rep(0.1, data$n_years_model + data$n_years_proj)
+  data$FMSY_init = rep(0.1, data$n_years_model + data$n_years_proj)
+  data$F_proj_init = 0
 
   # -------------------------------------------------------------------
   # Parameters