+++ src/helper_functions.hpp
@@ -1274,11 +1274,12 @@ Type get_SSB(matrix<Type> NAA, matrix<Type> ZAA, array<Type> waa, matrix<Type> m
 template <class Type>
 matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type> FAA, matrix<Type> NAA, matrix<Type> MAA, matrix<Type> mature, 
   vector<Type> waacatch, vector<Type> waassb, vector<Type> fracyr_SSB, vector<Type> log_SPR0, vector<int> avg_years_ind, 
-  int n_years_model, vector<int> which_F_age, Type percentSPR, vector<Type> proj_Fcatch, Type percentFXSPR, Type F_init){
+  int n_years_model, vector<int> which_F_age, Type percentSPR, vector<Type> proj_Fcatch, Type percentFXSPR, Type F_init, 
+  vector<Type> log_a, vector<Type> log_b, int recruit_model, Type percentFMSY){
     /* 
      get F to project for next time step
               y:  year of projection (>n_years_model)
-     proj_F_opt:  for each year, how to specify F for projection (1 to 5)
+     proj_F_opt:  for each year, how to specify F for projection (1 to 6)
         FAA_tot:  FAA_tot matrix from main code.
             NAA:  NAA matrix from main code
             MAA:  MAA matrix from main code.
@@ -1291,9 +1292,13 @@ matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type>
   n_years_model:  number of years before projection begins
     which_F_age:  define which age has max F
      percentSPR:  percentage (0-100) of unfished spawning potential to determine F_percentSPR
-    proj_Fcatch: vector (n_years_proj) of user specified Fishing mortality rates to project
+    proj_Fcatch:  vector (n_years_proj) of user specified Fishing mortality rates to project
    percentFXSPR:  percentage (0-100) of F_percentSPR to use in catch, e.g. GOM cod uses F = 75% F_40%SPR
-         F_init: initial value to use for FXSPR or FMSY newton method
+         F_init:  initial value to use for FXSPR or FMSY newton method
+          log_a:  annual log(a) for stock-recruit relationship
+          log_b:  annual log(b) for stock-recruit relationship
+  recruit_model:  integer for which type of recruit model is assumed (= 3 or 4 for using Fmsy)
+    percentFMSY:  percentage (0-100) of FMSY to use in catch.
     */
     //if(y > n_years_model-1){
     //for(int a = 0; a < n_ages; a++) waacatch(a) = waa(waa_pointer_totcatch-1, y, a);
@@ -1309,7 +1314,7 @@ matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type>
   if(proj_F_opt_y == 1){ // last year F (default)
     for(int f = 0; f < n_fleets; f++) for(int a = 0; a < n_ages; a++) FAA_proj(f,a) = FAA(n_years_model-1,f,a);
   }
-  if(proj_F_opt_y>1)
+  if(proj_F_opt_y==2)
   {
     //array<Type> FAA_toavg(n_fleets,n_toavg, n_ages);
     FAA_proj.setZero();
@@ -1322,6 +1327,14 @@ matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type>
       }
       FAA_proj.row(f) /= Type(n_toavg);
     }
+      
+    //proj_F_opt_y == 2, average F
+    if(proj_F_opt_y == 2){
+      //already defined
+      //for(int f = 0; f < n_fleets; f++) FAA_proj = FAA_toavg.colwise().mean();
+    }
+  }
+  if(proj_F_opt_y > 2){
     //get selectivity using average over avg.yrs
     matrix<Type> sel_proj(n_fleets,n_ages);
     vector<Type> sel_tot_proj(n_ages);
@@ -1332,19 +1345,14 @@ matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type>
     vector<Type> M = MAA.row(y);
     vector<Type> NAA_y = NAA.row(y);
     vector<Type> mat = mature.row(y);
-    
-    //proj_F_opt_y == 2, average F
-    if(proj_F_opt_y == 2){
-      //already defined
-      //for(int f = 0; f < n_fleets; f++) FAA_proj = FAA_toavg.colwise().mean();
-    }
-
+    Type fracyr_SSB_y = fracyr_SSB(y);
+    Type log_SPR0_y = log_SPR0(y);
+    Type log_a_y = log_a(y);
+    Type log_b_y = log_b(y);
     if(proj_F_opt_y == 3){ // F at X% SPR
-      Type fracyr_SSB_y = fracyr_SSB(y);
-      Type log_SPR0_y = log_SPR0(y);
       //FAA_tot.row(y) = get_FXSPR(M, sel_proj, waacatch, waassb, mat, percentSPR, fracyr_SSB_y, log_SPR0_y) * sel_proj;
-      FAA_tot_proj = get_FXSPR(M, sel_tot_proj, waassb, mat, percentSPR, fracyr_SSB_y, log_SPR0_y, F_init) * sel_tot_proj;
-      FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj * 0.01*percentFXSPR;
+      FAA_tot_proj = get_FXSPR(M, sel_tot_proj, waassb, mat, percentSPR, fracyr_SSB_y, log_SPR0_y, F_init) * sel_tot_proj * 0.01*percentFXSPR;
+      //FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj;
     }
     if(proj_F_opt_y == 4){ // user-specified F
       if(proj_Fcatch(y-n_years_model) < 1e-10){ // if F = 0, sel_proj is NaN
@@ -1352,7 +1360,7 @@ matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type>
       } else {
         //FAA_tot.row(y) = Type(proj_Fcatch(y-n_years_model)) * sel_proj;
         FAA_tot_proj = Type(proj_Fcatch(y-n_years_model)) * sel_tot_proj;
-        FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj;
+        //FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj;
       }
     }
     if(proj_F_opt_y == 5){ // calculate F from user-specified catch
@@ -1363,9 +1371,13 @@ matrix<Type> get_F_proj(int y, int n_fleets, vector<int> proj_F_opt, array<Type>
         //FAA_tot.row(y) = get_F_from_Catch(thecatch, NAA_y, M, sel_proj, waacatch) * sel_proj;
         //FAA_tot_proj = get_F_from_Catch(thecatch, NAA_y, M, sel_tot_proj, waacatch) * sel_tot_proj;
         FAA_tot_proj = get_F_from_log_Catch(thecatch, NAA_y, M, sel_tot_proj, waacatch, F_init) * sel_tot_proj;
-        FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj;
+        //FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj;
       }
     }
+    if(proj_F_opt_y == 6){ //Fmsy
+      FAA_tot_proj = get_FMSY(log_a_y, log_b_y, M, sel_tot_proj, waacatch, waassb, mat, fracyr_SSB_y, log_SPR0_y, recruit_model, F_init) * sel_tot_proj * 0.01* percentFMSY;
+    }
+    FAA_proj = FAA_tot_proj(which_F_age(y)-1) * sel_proj;
   }
   return(FAA_proj);
 }
@@ -1376,7 +1388,7 @@ matrix<Type> sim_pop(array<Type> NAA_devs, int recruit_model, vector<Type> mean_
   vector<Type> log_SR_b, vector<int> Ecov_where, vector<int> Ecov_how, matrix<Type> Ecov_lm, int n_NAA_sigma, 
   int do_proj, vector<int> proj_F_opt, array<Type> FAA, matrix<Type> FAA_tot, matrix<Type> MAA, matrix<Type> mature, array<Type> waa, 
   int waa_pointer_totcatch, int waa_pointer_ssb, vector<Type> fracyr_SSB, vector<Type> log_SPR0, vector<int> avg_years_ind, 
-  int n_years_model, int n_fleets, vector<int> which_F_age, Type percentSPR, vector<Type> proj_Fcatch, Type percentFXSPR, vector<Type> F_proj_init){
+  int n_years_model, int n_fleets, vector<int> which_F_age, Type percentSPR, vector<Type> proj_Fcatch, Type percentFXSPR, vector<Type> F_proj_init, Type percentFMSY){
 
   // Population model (get NAA, numbers-at-age, for all years)
   int ny = log_SR_a.size();
@@ -1416,7 +1428,7 @@ matrix<Type> sim_pop(array<Type> NAA_devs, int recruit_model, vector<Type> mean_
         vector<Type> waacatch = get_waa_y(waa, y, n_ages, waa_pointer_totcatch);
         vector<Type> waassb = get_waa_y(waa, y, n_ages, waa_pointer_ssb);
         matrix<Type> FAA_proj = get_F_proj(y, n_fleets, proj_F_opt, FAA, NAA, MAA, mature, waacatch, waassb, fracyr_SSB, log_SPR0, avg_years_ind, n_years_model,
-         which_F_age, percentSPR, proj_Fcatch, percentFXSPR, F_proj_init(y-n_years_model));
+         which_F_age, percentSPR, proj_Fcatch, percentFXSPR, F_proj_init(y-n_years_model), log_SR_a, log_SR_b, recruit_model, percentFMSY);
         for(int f = 0; f < n_fleets; f++) for(int a = 0; a< n_ages; a++) FAA(y,f,a) = FAA_proj(f,a);
         FAA_tot.row(y) = FAA_proj.colwise().sum();
         ZAA.row(y) = FAA_tot.row(y) + MAA.row(y);